pyproject.toml
[project]
name = "02-agent-output-type"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.3.2",
    "python-dotenv>=1.1.1",
]
